@@ -5,14 +5,14 @@ plugins {
 android {
 
     namespace 'com.wisethan.bestrefur1'
-    compileSdk 33
+    compileSdk 34
 
     defaultConfig {
         applicationId "com.wisethan.bestrefur1"
         minSdk 30
-        targetSdk 33
+        targetSdk 34
         versionCode 1
-        versionName "0.0.1"
+        versionName "0.0.3"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
@@ -47,6 +47,7 @@ dependencies {
 
     implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
     // OkHttp 라이브러리 추가 (Retrofit의 기본 HTTP 클라이언트)
+    implementation 'com.google.code.gson:gson:2.8.9'
     implementation 'com.squareup.okhttp3:okhttp:4.9.2'
     implementation 'com.davemorrissey.labs:subsampling-scale-image-view:3.10.0'
     //Glide
