@@ -1,17 +1,10 @@
 package com.wisethan.bestrefur1.common;
 
-import android.animation.ObjectAnimator;
-import android.animation.PropertyValuesHolder;
-import android.animation.ValueAnimator;
 import android.annotation.SuppressLint;
 import android.content.Intent;
-import android.graphics.Color;
 import android.os.Bundle;
 import android.os.Handler;
-import android.view.View;
-import android.view.animation.AccelerateDecelerateInterpolator;
 
-import androidx.annotation.RequiresPermission;
 import androidx.appcompat.app.ActionBar;
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -22,18 +15,14 @@ public class SplashActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         ActionBar actionBar = getSupportActionBar();
-        actionBar.hide();
-
+        if (actionBar != null) {
+            actionBar.hide();
+        }
         Handler handler = new Handler();
-        handler.postDelayed(new Runnable(
-
-        ) {
-        @Override
-        public void run() {
+        handler.postDelayed(() -> {
 
                 Intent intent = new Intent(getApplicationContext(), MainActivity.class);
                 startActivity(intent);finish();
-            }
-        }, 3000);//3초 있다가 메인 액티비티로
+            }, 3000);//3초 있다가 메인 액티비티로
     }
 }
