@@ -17,6 +17,7 @@ import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.bumptech.glide.Glide;
+import com.bumptech.glide.load.engine.DiskCacheStrategy;
 import com.wisethan.bestrefur1.R;
 import com.wisethan.bestrefur1.RebornOrder.model.Reborn;
 
@@ -29,13 +30,9 @@ public class TheRebornAdapter extends RecyclerView.Adapter<TheRebornAdapter.Rebo
     private final List<Reborn> rebornsList; // 리본 리스트
 
     private final List<Reborn> filteredRebornList; // 필터링된 상품명 리스트
-
-    @SuppressLint("NotifyDataSetChanged")
     public TheRebornAdapter(List<Reborn> rebornsList) {
-        setHasStableIds(true);
         this.rebornsList = rebornsList;
         this.filteredRebornList = new ArrayList<>(rebornsList);
-        notifyDataSetChanged();
     }
 
     @NonNull
@@ -51,10 +48,19 @@ public class TheRebornAdapter extends RecyclerView.Adapter<TheRebornAdapter.Rebo
 
         Glide.with(holder.itemView.getContext())
                 .load(reborn.getProductImgUrl())
+                //기본적으로 메모리 캐싱을 하기때문에, 메모리 캐싱을 위해 추가적으로 할 일은 없다.
+                //단, URL 이미지 로딩 시 한번 로드한 이미지는 chache 에 저장되어 서버에서 해당 이미지를 변경해도
+                //App 의 이미지는 갱신되지 않는다.
+                //이런 경우, skipMemoryCache(true) 로 메모리 캐시를 사용하지 않을 수 있어 skipMemoryCache()메서드를 사용.
+                // 디스크캐시 건너띄기 iskCacheStrategy(DiskCacheStrategy.NONE)
+                // 둘다 건너띄기는 같이 쓰면됨
+                // 일반적으로 캐시를 건너 뛰지 않는걸 권장.
+                // 이미지를 검색, 디코딩 및 변환하여 새 썸네일을 만드는 것보다 캐시에서 이미지를 load하는 것이 빠름
+                .skipMemoryCache(true)// 메모리 캐시 저장 off
+                .diskCacheStrategy(DiskCacheStrategy.ALL)// 디스크 캐시 저장
                 .into(holder.goods_view); // Glide를 사용하여 이미지 로드
 
-        holder.itemView.setLayerType(View.LAYER_TYPE_HARDWARE, null);
-
+        holder.goods_view.setLayerType(View.LAYER_TYPE_HARDWARE, null);
         holder.goods_name.setText(reborn.getGoodsName()); // 상품명 설정
 
         holder.itemView.setOnClickListener(v -> {
@@ -108,7 +114,7 @@ public class TheRebornAdapter extends RecyclerView.Adapter<TheRebornAdapter.Rebo
     public static class RebornViewHolder extends RecyclerView.ViewHolder {
 
         ImageView goods_view; // 상품 이미지를 보여주는 ImageView
-        TextView goods_name; // 상품명을 보여주는 TextView
+        final TextView goods_name; // 상품명을 보여주는 TextView
 
         public RebornViewHolder(@NonNull View itemView) {
             super(itemView);
